feat : SortedSection을 추가하여 역할 분리
